requirements.txt
google-adk
duckdb
python-dotenv